Add SASL authentication to CQL native protocol
Added new message types for Client/Server to perform SASL authentication.
Credentials messages should be deprecated (it hasn't been removed yet, but
is now unused).

SaslAuthBridge represents the interface between SASL & IAuthencator, and
the helper class org.apache.cassandra.transport.sasl.Sasl acts as a registry
of which SaslAuthBridge implementation goes with which IAuthenticator.
PasswordAuthenticator and other custom IAuthenticator implementations
which obtain credentials via Credentials message (or thrift login() call)
should be associated with PlainTextSaslAuthBridge. (This is done
automatically for PasswordAuthenticator).

Implementors of custom authenticators which do not receive credentials in
the standard way will need to provide their own SaslAuthBridge which can
extract the credentials to be used by the authenticator from a SaslServer.
This may involve creating or wrapping a SaslServer implementation.
AbstractSaslAuthBridge provides a potential base class for such
implementations.